Rename upload item component
