+++ client/src/components/ActivityBar/Items/UploadItem.vue
